Create Multilayer more complex
In an effort to make this work, I have done it the other way, however the loop for producing the R and T etc is now only producing a single value (the final value) and is no longer creating an array of values. Therefore the produced values cannot be plotted!

Currently not sure why, will investigate further!